# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/": {
            "post": {
                "description": "Сохраняет URL и возвращает его сокращенную версию",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "URL"
                ],
                "summary": "Сохранить URL в текстовом формате",
                "parameters": [
                    {
                        "description": "URL для сокращения",
                        "name": "url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Сокращенный URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Сокращенный URL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/shorten": {
            "post": {
                "description": "Сохраняет URL и возвращает его сокращенную версию",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL"
                ],
                "summary": "Сохранить URL в формате JSON",
                "parameters": [
                    {
                        "description": "URL для сокращения",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ShortenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.ShortenResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.ShortenResponse"
                        }
                    }
                }
            }
        },
        "/api/shorten/batch": {
            "post": {
                "description": "Сохраняет массив URL и возвращает массив сокращенных URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL"
                ],
                "summary": "Пакетное сохранение URL",
                "parameters": [
                    {
                        "description": "Массив URL для сокращения",
                        "name": "batch",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BatchRequestItem"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BatchResponseItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/urls": {
            "get": {
                "description": "Возвращает список всех URL, созданных пользователем",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL"
                ],
                "summary": "Получить все URL пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.UserURLItem"
                            }
                        }
                    },
                    "204": {
                        "description": "Нет сохраненных URL"
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет указанные URL пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "URL"
                ],
                "summary": "Удалить URL пользователя",
                "parameters": [
                    {
                        "description": "Массив идентификаторов URL для удаления",
                        "name": "urls",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Запрос на удаление принят"
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{id}": {
            "get": {
                "description": "Перенаправляет на оригинальный URL по сокращенному идентификатору",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "URL"
                ],
                "summary": "Получить оригинальный URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор сокращенного URL",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Редирект на оригинальный URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "410": {
                        "description": "URL был удален",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.BatchRequestItem": {
            "type": "object",
            "properties": {
                "correlation_id": {
                    "description": "Идентификатор корреляции для связи запроса и ответа",
                    "type": "string"
                },
                "original_url": {
                    "description": "Оригинальный URL для сокращения",
                    "type": "string"
                }
            }
        },
        "dto.BatchResponseItem": {
            "type": "object",
            "properties": {
                "correlation_id": {
                    "description": "Идентификатор корреляции из запроса",
                    "type": "string"
                },
                "short_url": {
                    "description": "Сокращенный URL",
                    "type": "string"
                }
            }
        },
        "dto.UserURLItem": {
            "type": "object",
            "properties": {
                "original_url": {
                    "description": "Оригинальный URL",
                    "type": "string"
                },
                "short_url": {
                    "description": "Сокращенный URL",
                    "type": "string"
                }
            }
        },
        "handlers.ShortenRequest": {
            "type": "object",
            "properties": {
                "url": {
                    "description": "URL для сокращения",
                    "type": "string",
                    "example": "https://example.com"
                }
            }
        },
        "handlers.ShortenResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "description": "Сокращенный URL",
                    "type": "string",
                    "example": "http://localhost:8080/abc123"
                }
            }
        }
    }
}